Parameterize some more aspects of AIs and adjust average AI
Don't keep sequences of three cards and only keep the highest card without at least a pair or a flush or straight draw.

This shouldn't make the average AI much harder; in fact, it may actually be easier to play against because it won't swap three cards all the time regardless of whether it has a pair or not.

Also remember that the average AI is used for the card suggest feature, and keeping the top two cards or sequences of three cards is generally obviously bad strategies.

I introduced these flaws to compensate for fixing bugs in the AI. It is still the case that the average AI goes for straights over keeping a pair.