Unlockable costumes
This feature allows a costume/outfit/reskin to be hidden until its  associated collectible is unlocked.

All of the code changes in player.js, spniDisplay.js, and spniSelect.js are by Erion and are identical to !9743 merged into the `editor68` branch last year.

Also part of this MR is an example of this in action: by adding a "collectible=" property to the applicable costume line in meta.xml, the connection is made. Check the draft with Nagisa's files and a new collectible with the following temporarily-easy unlock method: select Nagisa, select Komi-san, deselect Komi-san. Note that Nagisa does not have a visible costume before this, but does afterward.

Note that unlocking the collectible for a costume does *not* currently override seasonable visibility, as this was not Erion's goal when creating the feature. This is something we would have to add later to align with the v13 announcement. For the purposes of this demonstration, I have removed the Valentine's categorization for Nagisa's costume.